Change the way of storing command blocks to map
From JSONObject to HashMap, the reason I was doing this is because I can use Long as key and value at first time.
Update JDA version.
Now it pay respect on Mega.
Add Mandarin in online and offline message.
(Maybe) fixed a bug of dm new user.
Change the way of regex matching in OneATwoBGame.java to a more efficiency way.