The 100th commit
Replace HashMap declare to Map.
Change argument in /one_a_two_b to integer.
Make JsonHandle.builcCommandBlocksMap() return Map directly.
Add QuestionForumHandle.createForumPost being a method that executes when a question forum post was created.
Make OneATwoB.calculateAAndB return integer array instead of integer to avoid divide.
Remove CommandBlocksHandle.getMap(), replace with CommandBlocksHandle.getKeySet().
Fix a bug that if a user is not cached, they can't send message to underground through this bot because this bot think he/she is not a member of the server.