Update JDA to 5.0.0
Add reboot option that choose if the bot will send message to bot channel when online.
Use image file to send sticker from dm to underground.
Use id compare to detect if the message is the first message of a forum post.
Add poop map jam forum to ForumMessage.
Add sealed to ShowcaseThreadButtons.